fix short long and added descending order

+++ migrations/20200907113818-create-daily-swap.js
@@ -21,10 +21,10 @@ module.exports = {
         type: Sequelize.STRING
       },
       swap_long: {
-        type: Sequelize.INTEGER
+        type: Sequelize.DOUBLE
       },
       swap_short: {
-        type: Sequelize.INTEGER
+        type: Sequelize.DOUBLE
       },
       // createdAt: {
       //   allowNull: false,
